main_service/requirements.txt
fastapi
uvicorn
jinja2
httpx
python-multipart
uuid
pyjwt==2.10.1